+++ Fuji/Source/Drivers/PS2/MFSockets_PS2.cpp
@@ -43,22 +43,22 @@ MFSocket MFSockets_Accept(MFSocket socket, MFSocketAddress *pConnectingSocketAdd
 	return 0;
 }
 
-int MFSockets_Send(MFSocket socket, const char *pBuffer, int bufferLength, uint32 flags)
+int MFSockets_Send(MFSocket socket, const char *pBuffer, size_t bufferLength, uint32 flags)
 {
 	return bufferLength;
 }
 
-int MFSockets_SendTo(MFSocket socket, const char *pBuffer, int bufferLength, uint32 flags, const MFSocketAddress *pAddress)
+int MFSockets_SendTo(MFSocket socket, const char *pBuffer, size_t bufferLength, uint32 flags, const MFSocketAddress *pAddress)
 {
 	return bufferLength;
 }
 
-int MFSockets_Recv(MFSocket socket, char *pBuffer, int bufferSize, uint32 flags)
+int MFSockets_Recv(MFSocket socket, char *pBuffer, size_t bufferSize, uint32 flags)
 {
 	return bufferSize;
 }
 
-int MFSockets_RecvFrom(MFSocket socket, char *pBuffer, int bufferSize, uint32 flags, MFSocketAddress *pSenderAddress)
+int MFSockets_RecvFrom(MFSocket socket, char *pBuffer, size_t bufferSize, uint32 flags, MFSocketAddress *pSenderAddress)
 {
 	return bufferSize;
 }
