+++ Fuji/Source/Drivers/WIN32/MFSockets_WINSOCK.cpp
@@ -255,12 +255,12 @@ MFSocket MFSockets_Accept(MFSocket socket, MFSocketAddress *pConnectingSocketAdd
 	return result;
 }
 
-int MFSockets_Send(MFSocket socket, const char *pBuffer, int bufferLength, uint32 flags)
+int MFSockets_Send(MFSocket socket, const char *pBuffer, size_t bufferLength, uint32 flags)
 {
-	return send((SOCKET)socket, pBuffer, bufferLength, flags);
+	return send((SOCKET)socket, pBuffer, (int)bufferLength, flags);
 }
 
-int MFSockets_SendTo(MFSocket socket, const char *pBuffer, int bufferLength, uint32 flags, const MFSocketAddress *pAddress)
+int MFSockets_SendTo(MFSocket socket, const char *pBuffer, size_t bufferLength, uint32 flags, const MFSocketAddress *pAddress)
 {
 	int addrLen = 0;
 	sockaddr *pSockAddr = NULL;
@@ -271,20 +271,20 @@ int MFSockets_SendTo(MFSocket socket, const char *pBuffer, int bufferLength, uin
 		MFDebug_Assert(pSockAddr, "Invalid socket address...");
 	}
 
-	return sendto((SOCKET)socket, pBuffer, bufferLength, flags, pSockAddr, addrLen);
+	return sendto((SOCKET)socket, pBuffer, (int)bufferLength, flags, pSockAddr, addrLen);
 }
 
-int MFSockets_Recv(MFSocket socket, char *pBuffer, int bufferSize, uint32 flags)
+int MFSockets_Recv(MFSocket socket, char *pBuffer, size_t bufferSize, uint32 flags)
 {
-	return recv((SOCKET)socket, pBuffer, bufferSize, flags);
+	return recv((SOCKET)socket, pBuffer, (int)bufferSize, flags);
 }
 
-int MFSockets_RecvFrom(MFSocket socket, char *pBuffer, int bufferSize, uint32 flags, MFSocketAddress *pSenderAddress)
+int MFSockets_RecvFrom(MFSocket socket, char *pBuffer, size_t bufferSize, uint32 flags, MFSocketAddress *pSenderAddress)
 {
 	char address[sizeof(sockaddr_in)*10];
 	int size = sizeof(sockaddr_in)*10;
 
-	int result = recvfrom((SOCKET)socket, pBuffer, bufferSize, flags, (sockaddr*)&address, &size);
+	int result = recvfrom((SOCKET)socket, pBuffer, (int)bufferSize, flags, (sockaddr*)&address, &size);
 
 	if(pSenderAddress)
 	{
